+++ tests/compilers/test_ast.py
@@ -396,8 +396,6 @@ def test_ast_non_decoratable():
 
 def test_ast_lambda_lists():
     """Ensure the compiler chokes on invalid lambda-lists"""
-    cant_compile('(fn [&key {"a" b} &key {"foo" bar}] [a foo])')
-    cant_compile('(fn [&optional a &key {"foo" bar}] [a foo])')
     cant_compile('(fn [&optional [a b c]] a)')
     cant_compile('(fn [&optional [1 2]] (list 1 2))')
 
@@ -425,13 +423,6 @@ def test_lambda_list_keywords_rest():
     can_compile("(fn (x &rest xs) (print xs))")
     cant_compile("(fn (x &rest xs &rest ys) (print xs))")
     can_compile("(fn (&optional a &rest xs) (print xs))")
-
-
-def test_lambda_list_keywords_key():
-    """ Ensure we can compile functions with &key."""
-    can_compile("(fn (x &key {foo True}) (list x foo))")
-    cant_compile("(fn (x &key {bar \"baz\"} &key {foo 42}) (list x bar foo))")
-    cant_compile("(fn (x &key {1 2 3 4}) (list x))")
 
 
 def test_lambda_list_keywords_kwargs():